refactor : clean package and remove setter method of User class
