+++ src/main/java/techcourse/myblog/controller/ArticleController.java
@@ -1,10 +1,10 @@
-package techcourse.myblog.web;
+package techcourse.myblog.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.model.Article;
 import techcourse.myblog.repository.ArticleRepository;
 
 @Controller
