+++ src/main/java/techcourse/myblog/controller/dto/UserDTO.java
@@ -1,0 +1,13 @@
+package techcourse.myblog.controller.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class UserDTO {
+    private String userName;
+    private String email;
+    private String password;
+
+}
