+++ src/main/java/techcourse/myblog/model/User.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.model;
 
 import lombok.*;
 
@@ -8,7 +8,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Getter
-@Setter
 @NoArgsConstructor
 @RequiredArgsConstructor
 @Entity
