+++ src/main/java/techcourse/myblog/repository/ArticleRepository.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.repository;
 
 import org.springframework.stereotype.Repository;
-import techcourse.myblog.domain.Article;
+import techcourse.myblog.model.Article;
 
 import java.util.ArrayList;
 import java.util.List;
