+++ src/main/java/techcourse/myblog/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import techcourse.myblog.domain.User;
+import techcourse.myblog.model.User;
 
 public interface UserRepository extends CrudRepository<User, Long> {
 }
