+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -1,0 +1,5 @@
+package techcourse.myblog.service;
+
+public class UserService {
+
+}
